moved exceptions to utils
